[functions] BACKPORT Allow functions to pass runtime specific options
This commit adds a new argument for functions, customRuntimeOptions, which is passed to
funcions (as well as sources/sinks) that enables the ability to
customize the runtime.

This is added primarily to support the `KubernetesManifestCustomizer`
interface. This interface has a default implementation, which allows for
changing the namespace, labels, and annotations per function.

This interface is also pluggable, allowing for more customizable
implementations. For example, the functions for a given tenant could be
mapped to different pools of compute for isolation.