Fix bug where a producer get incorrectly cached
When an attempt to create a producer fails due
to the topic being owned by a different instance
the producer is still cached.

If the instance later gains ownership of the topic
the client won't be able to create a producer due
to the previously cached producer.

This prevents the producer from being cached if the
broker doesn't own the topic.