[functions] Provide more options for func worker prometheus metrics
The broker and proxy both allow for hitting the metrics endpoint without auth. The functions
worker should allow that to be configurable as well. This adds an option to
allow for metrics endpoint to allow the endpoint to be hit without auth

Additionally, the functions worker doesn't expose the default prometheus
metrics (such as JVM info, etc).

This commit implements and adds an option to support that